backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pinecone-client==3.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
google-generativeai==0.4.0
pydantic==2.6.3
pydantic-settings==2.2.1
aiofiles==23.2.1
requests==2.31.0
typing-extensions==4.10.0
